Import encounters as separate assets
This will allow us to assign different encounters to different fields, so that we can handle the case at the end of having to choose between different encounters for the last one.  It also makes it possible to create and configure additional encounters directly in unity by manually creating new Encounter assets via the context menu.